+++ src/modules/users/edit-user-sheet/api/edit-user-sheet-api.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { Usuario } from './edit.api-model';
+import { Usuario } from './edit-user-sheet-api-model';
 
 export const getUserById = async (id: string): Promise<Usuario> => {
   const response = await axios.get<Usuario>(`/api/user/${id}`);
