+++ src/modules/users/edit-user-sheet/api/index.ts
@@ -1,2 +1,2 @@
-export * from './edit-api';
-export * from './edit.api-model';
+export * from './edit-user-sheet-api';
+export * from './edit-user-sheet-api-model';
