+++ src/modules/users/edit-user-sheet/edit-user-sheet.query.hook.ts
@@ -1,8 +1,8 @@
 import { useNavigate } from '@tanstack/react-router';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { queryClient, usersQueryKeys } from '#core/react-query';
-import { getUserByIdRepository, updateUser } from './edit.repository';
-import { createEmptyUsuario, Usuario } from './edit.vm';
+import { getUserByIdRepository, updateUser } from './edit-user-sheet.repository';
+import { createEmptyUsuario, Usuario } from './edit-user-sheet.vm';
 
 interface UseSaveUserMutationResult {
   saveUser: (user: Usuario) => void;
